fix: Scroll does not reset when navigating shared docs on mobile
closes #6968
