+++ app/index.tsx
@@ -20,7 +20,7 @@ import env from "~/env";
 import { initI18n } from "~/utils/i18n";
 import Desktop from "./components/DesktopEventHandler";
 import LazyPolyfill from "./components/LazyPolyfills";
-import MobileScrollWrapper from "./components/MobileScrollWrapper";
+import PageScroll from "./components/PageScroll";
 import Routes from "./routes";
 import Logger from "./utils/Logger";
 import history from "./utils/history";
@@ -61,7 +61,7 @@ if (element) {
                   <LazyPolyfill>
                     <LazyMotion features={loadFeatures}>
                       <Router history={history}>
-                        <MobileScrollWrapper>
+                        <PageScroll>
                           <PageTheme />
                           <ScrollToTop>
                             <Routes />
@@ -69,7 +69,7 @@ if (element) {
                           <Toasts />
                           <Dialogs />
                           <Desktop />
-                        </MobileScrollWrapper>
+                        </PageScroll>
                       </Router>
                     </LazyMotion>
                   </LazyPolyfill>
